+++ src/posts/posts.service.ts
@@ -1,8 +1,8 @@
 
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PostEntity } from './entities/post.entity';
-import { DeleteResult, Repository, UpdateResult } from 'typeorm';
+import { Repository } from 'typeorm';
 
 @Injectable()
 export class PostsService {
@@ -17,8 +17,8 @@ export class PostsService {
     });
   }
 
-  findOne(id: number): Promise<PostEntity> {
-    return this.postsRepository.findOne({
+  async findOne(id: number): Promise<PostEntity> {
+    return await this.postsRepository.findOne({
       where: { id },
       relations: ["user_id"]
     });
@@ -28,12 +28,72 @@ export class PostsService {
     return this.postsRepository.save(post);
   }
 
-  remove(id: number): Promise<DeleteResult> {
-    return this.postsRepository.delete(id);
+  async remove(id: number, user_id: number): Promise<HttpStatus> {
+    const post:PostEntity = await this.findOne(id);
+        
+    if(post?.user_id?.id == user_id)
+    {
+      await this.postsRepository.delete(id);
+      return HttpStatus.OK
+    }      
+    else
+      throw new HttpException('Você não tem permissão ou este Post não existe', HttpStatus.NOT_FOUND);      
   }
 
-  updateDescription(id: number, post: PostEntity): Promise<PostEntity> {
+  async updateDescription(id: number, post: PostEntity, user_id: number): Promise<PostEntity> {
+    const current_post: PostEntity = await this.findOne(id)
+    
+    if(current_post?.user_id?.id == user_id)
+    {
+      current_post.edition_history.push({ title: current_post.title, description: current_post.description })
+      post.edition_history = current_post.edition_history
+      this.postsRepository.update(id, post);
+      return this.findOne(id)
+    }
+    else
+      throw new HttpException('Você não tem permissão ou este Post não existe', HttpStatus.NOT_FOUND);
+  }
+
+  async like(post_id: number, user_id: number): Promise<PostEntity> {
+    const post: PostEntity = await this.findOne(post_id)
+
+    if (post.disliked_by.find((value) => value === user_id))
+    {
+      post.disliked_by = post.disliked_by.filter((value) => value != user_id)
+      post.dislikes--;
+    }
+
+    if (!post.liked_by.find((value) => value === user_id))          
+    {
+      post.liked_by.push(user_id)
+      post.likes++;
+      this.postsRepository.update(post_id, post);
+    }
+    return post
+  }
+
+  async dislike(post_id: number, user_id: number): Promise<PostEntity> {
+    const post: PostEntity = await this.findOne(post_id)
+
+    if (post.liked_by.find((value) => value === user_id))      
+    {
+      post.liked_by = post.liked_by.filter((value) => value != user_id)
+      post.likes--;
+    }
+
+    if (!post.disliked_by.find((value) => value === user_id))      
+    {    
+      post.disliked_by.push(user_id)
+      post.dislikes++;
+      this.postsRepository.update(post_id, post);
+    }    
+    return post
+  }
+
+  async viewed(id: number): Promise<PostEntity> {
+    const post: PostEntity = await this.findOne(id)
+    post.views++;
     this.postsRepository.update(id, post);
-    return this.findOne(id)
+    return this.findOne(id);
   }
 }
